pyproject.toml
[tool.poetry]
name = "playwright-python"
version = "0.1.0"
description = ""
authors = ["Carlos Kidman <carlos@qap.dev>"]

[tool.poetry.dependencies]
python = "^3.10"
typer = { extras = ["all"], version = "^0.7.0" }
python-dotenv = "^1.0.0"
pytest = "^7.2.2"
pytest-playwright = "^0.3.2"
pytest-xdist = "^3.2.1"

[tool.poetry.dev-dependencies]
black = "^23.1.0"
flake8 = "^6.0.0"
ipykernel = "^6.21.3"

[tool.black]
line-length = 120

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
